ml_service/requirements.txt
flask
flask-cors
ultralytics
opencv-python
pillow
numpy
torch
torchvision
